refactor: extract error responder to backend/pkg/respond
